add confirmation on delete; fix delete on some endpoints; update with soft delete permit types, qualities, product types

+++ resources/js/Components/Management/AAC/Parts/InventoryGrid/Vue/ActionColumn.vue
@@ -16,6 +16,7 @@
 import Switches from 'vue-switches';
 import AAC from "components/Management/AAC/AAC";
 import Notification from "components/Common/Notifications/Notification";
+import Confirmation from "../../../../../Common/Confirmation/Confirmation";
 
 export default {
     props: ["rowProp", "optionsProp"],
@@ -24,9 +25,14 @@ export default {
 
     methods: {
         deleteInventory () {
-            AAC.delete_inventory(this.rowProp.id).finally(() => {
-                Vent.$emit('grid-refresh')
+            return Confirmation(this.translate('corfirmation_delete_question')).then((result) => {
+                if(result.isConfirmed) {
+                    AAC.delete_inventory(this.rowProp.id).finally(() => {
+                        Vent.$emit('grid-refresh')
+                    })
+                }
             })
+
         },
         approve(value) {
             AAC.approveInventory(this.rowProp.Id, {Approved: value}).then((response) => {
